fix: rename files to be different than containing directory
Prevents "Maximum call stack size exceeded" when building with
"flatModuleOutFile".
angular#25226
