+++ src/framework/theme/components/cdk/bidi/bidi.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { BidiModule, Directionality } from '@angular/cdk/bidi';
-import { NbDirectionality } from './bidi';
+import { NbDirectionality } from './bidi-service';
 
 @NgModule({
   providers: [
