+++ src/framework/theme/components/cdk/overlay/index.ts
@@ -1,5 +1,5 @@
 export * from './overlay.module';
-export * from './overlay';
+export * from './overlay-service';
 export * from './overlay-position';
 export * from './overlay-container';
 export * from './overlay-trigger';
