+++ src/framework/theme/components/cdk/platform/platform-service.ts
