+++ src/framework/theme/components/cdk/platform/platform.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Platform, PlatformModule } from '@angular/cdk/platform';
-import { NbPlatform } from './platform';
+import { NbPlatform } from './platform-service';
 
 @NgModule({
   providers: [
